pyproject.toml
[project]
name = "fastapi-blueprint"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "fastapi[standard]>=0.112.1",
    "uvicorn==0.30.5",
    "pydantic-settings==2.3.4"
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.8.4",
    "pre-commit~=4.2.0"
]
